requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
six==1.12.0
Werkzeug==0.16.0
Flask-SocketIO==4.2.1
python-dotenv==0.10.3
APScheduler==3.6.3
pyserial==3.4
wifi==0.3.8
